Refactor: 소켓기능을 socket service로 변경 - socket 변수를 즉시실행함수로 url, options을 받아 실행하도록 변경 - socket 객체는 init, disconnect, isDisconnect 메서드를 반환
